Define wriggle behavior as spec
Create wriggle behavior spec types that are used to generate wriggle functions instead of defining the funcitons directly.

This is necessary because ultimately Pals will be defined as json, so any behavior functions need to be derived from basic data.

Also rename Wiggle to Wriggle, which I think is more accurate. Filename change left for follow-up commit so git registers it as a rename, not add/delete.