Specify sizes in Pal Units until render
Set a top-level size for Pals (hard-coded for now). All segment and movement logic should deal with sizes in Pal Units: aka as a percent of this top-level size.

This way, we can (mag/min)ify Pal by adjusting the top-level size, and all movements and segments will stay proportional.