Tighten section paradigm: couple spec/segmentation
Remove generalized functions for specific section creation.
The point of the segmentation paradigm is to have self-contained nodes we can compose into a tree. Where possible, these nodes should define their own spec relative to their parent, instead of having their parent define its child's spec relative to it.
This means that, when it comes to the chimera pal that is somewhat randomized, we can mix-and-match child and parent nodes without needing to understand how to parameterize the children too much.