Cleanup: move vite config and html into src
Pull html out of root /client directory. This is mostly cosmetic, but it is becoming confusing/messy having various html entrypoints in the root /client dir.

Ideally *.html would go in a dedicated public/entry subdir, but vite really needs it to be in a parent of src (or at best in src itself).

There's sort of a Catch 22 between the vite dev and vite build requirements:
* vite dev needs to run from the root (or src) directory, so our html can import the files in src (it cannot access files outside of root)
* vite build need to run from the same directory as our html, otherwise the html will be written to a sub-directory within dist

So the only way to not have html in the root client dir (without really fighting vite) is to have both the html and vite in src. Now dev and build are happy.