Fix: pal modifier bug in Exhibit component
The `mod` field is now set correctly (was not updated when field name changed from `settings`).
A magnification bug is also fixed. Before, magnification directly influcenced the segment render size. Now it is a factor that affects the pal size field. The modifier component now sets magnification between 20 and 500, rather than 2 and 50, to compensate.