Fix: container/visibility hook race case.
Pass the full div ref as container context, rather than just the value. This will allow the intersection observer to reference the correct element even without a ref.

This steps outside the React paradigm, but I think it's ok in this case for essentially an animation effect that doesn't fit well into React's purpose & expectations.